requirements.txt
langchain
langchain_community
arxiv
wikipedia
langchain-huggingface
langchain-openai
faiss-cpu
ipykernel
python-dotenv
langchain-groq
streamlit
duckduckgo-search